--- profile_maker/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ProfileMakerConfig(AppConfig):
-    name = 'profile_maker'
