--- profile_maker/templates/profile_maker/details.html
@@ -1,24 +1,0 @@
-{#Source: https://data-flair.training/blogs/django-file-upload/#}
-{#Author: Laura Timm#}
-{#Date Created: March 29, 2020#}
-{#This extends file html page, used when the file is successfully uploaded #}
-
-{% extends 'profile_maker/create.html' %}
-
-{% block replace %}
-<CENTER>
-<div class="jumbotron">
-    <h2 class="display-4">Thank You {{ user_pr.fname }}<br> </h2>
-  <p class="lead">Your submission has been uploaded successfully<br></p>
-  <hr class="my-4">
-    <object data="{{ user_pr.display_picture.url }}" type="application/pdf" width="100%" height="100%">
-  <p>Alternative text - include a link <a href="{{ user_pr.display_picture.url }}">to the PDF!</a></p>
-</object>
-{#    <embed src="{{user_pr.display_picture.url }}" type="application/pdf" height="700px" width="500"> <br>#}
-
-
-    {{ user_pr.display_picture.url }}<BR>
-   You have learned some awesome technologies like {{ user_pr.technologies }}</p>
-</div>
-</CENTER>
-{% endblock %}
