+++ userPages/urls.py
@@ -1,6 +1,6 @@
+from django.conf.urls import include, url
 from django.urls import path
 from . import views
-from django.conf.urls import include, url
 
 from website import settings
 from django.conf.urls.static import static
@@ -11,4 +11,10 @@ urlpatterns = [
 
     path('proposal_list/', views.reviewer_view_proposals, name='Proposal View'),
     path('proposal/<int:pk>', views.ProposalDetailView.as_view(), name='proposal-detail'),
+    path('', views.create_profile, name = 'create'),
+
 ]
+
+if settings.DEBUG:
+    urlpatterns += static(settings.STATIC_URL, document_root = settings.STATIC_ROOT)
+    urlpatterns += static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT)
