+++ website/urls.py
@@ -26,9 +26,7 @@ urlpatterns = [
 
     path('login/', views.loginRequest),
     path('', views.home),
-    path('upload/', include('profile_maker.urls')),
     path('', include('userPages.urls')),
-    path('', include('userPages.urls')),
-    path('', include('userPages.urls')),
+
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
